Replace incorrect dummy size fixup in compressed BinaryBuffer
If `abWriteDataSize` is set, a zero value is written, followed by the
actual compressed data, and the zero value is then replaced with the
actual size value.
It looks like the original code meant to `mpData + lStartPos`, but
forgot the address-of operation in the expression
`(int*)mpData[lStartPos]`, which instead takes the (char) value
`mpData[lStartPos]` and uses that as address to write to.
This write is replaced by a `memcpy()`, as used in other places to write
to the buffer, which not only fixes the aforementioned issue, but also
avoids pointer alignment and aliasing issues.