A few small build and runtime fixes (#8)
* Fix out-of-bounds access in cEditorWindowViewport

Various editors create viewports with a "View"-menu, from which a render
mode can be selected.
The engine provides three render modes: Shaded, WireFrame and Simple.
The editor viewports only provide menu options for the first two, while
one loop tried to iterate over an array of render modes by checking
against the eRenderer_LastEnum value, which is 3 to include all three
modes.
Since the array in question, mpMainMenuRenderModes, only provides the
first two modes, the last iteration would attempt to access a third
element, which lies out of bounds.
This is fixed by replacing the check against a constant literal value of
"2", thusly correctly stopping after checking the first two elements of
the array.

* Fix C++11 deprecation warning

std::binary_function is deprecated in C++11 and removed in later
versions. Its use can simply be omitted.

* Replace incorrect dummy size fixup in compressed BinaryBuffer

If `abWriteDataSize` is set, a zero value is written, followed by the
actual compressed data, and the zero value is then replaced with the
actual size value.
It looks like the original code meant to `mpData + lStartPos`, but
forgot the address-of operation in the expression
`(int*)mpData[lStartPos]`, which instead takes the (char) value
`mpData[lStartPos]` and uses that as address to write to.
This write is replaced by a `memcpy()`, as used in other places to write
to the buffer, which not only fixes the aforementioned issue, but also
avoids pointer alignment and aliasing issues.

* Correct cSurfacePicker::SetAverageNormal() function signature

* Fix build for USE_GAMEPAD=OFF variant

The USE_GAMEPAD builds variable controls gamepad support.
Various pieces of code expect the `cLuxInputHandler::IsGamepadPresent()`
method to be available unconditionally.
This is fixed by simply providing the method for USE_GAMEPAD=OFF builds
now too, simply always returning false.
Some conditionally compiled code using `#ifdef USE_GAMEPAD` is
simplified accordingly.

* Fix runtime errors for USE_GAMEPAD=OFF variant

The lean action is also valid and available for non-gamepad use.
It is bound to left Alt by default and uses horizontal mouse movement to
change the lean angle.

* Reenable ParticleViewer

* Reenable LevelEditor